use getTargetValue for getColor functions in LedStrip
This is because whoever is using this function probably won't be expecting the exact color at that moment,
but we're also not returning the currentColor and instead recreating it. This makes it more failSafe and it always gives correct results (except while is changing). all these has to be documented